+++ frontend/src/components/UI/AdminMenu.vue
@@ -1,8 +1,9 @@
 <template>
   <b-col cols="12" md="4" lg="2" class="margin-top">
     <span class="font-weight-bold">Administracja</span><br/>
-    <a href="" @click.prevent="$router.push({name: 'PetAdmin'})">Zwierzęta</a><br/>
-    <a href="" @click.prevent="$router.push({name: 'PetTypesAdmin'})">Typy zwierząt</a><br/>
+    <a href="" @click.prevent="$router.push({name: 'PetAdmin'})" :class="{active: $route.name==='PetAdmin'}">Zwierzęta</a><br/>
+    <a href="" @click.prevent="$router.push({name: 'PetTypesAdmin'})" :class="{active: $route.name==='PetTypesAdmin'}">Typy zwierząt</a><br/>
+    <a href="" @click.prevent="$router.push({name: 'ActionTypesAdmin'})" :class="{active: $route.name==='ActionTypesAdmin'}">Typy aktywności</a><br/>
     <a href="">Schroniska</a><br/>
     <a href="">Aktualności</a><br/>
     <a href="">Użytkownicy</a><br/>
@@ -27,5 +28,9 @@ export default {
 }
 a {
   color: #42b983;
+}
+
+a.active{
+  color: #286f4f;
 }
 </style>