+++ CHANGELOG.md
@@ -5,5 +5,7 @@
 - Lista zwierząt (/pets)
 - Lista schronisk (/shelters)
 - Administracyjna lista zwierząt (/admin/pets) + przyciski do narzędzi
+- Administracyjna lista typów zwierząt (/admin/pet-types)
+- Administracyjna lista typów aktywności (/admin/action-types)
 - Menu administracyjne